+++ src/Firebase/Value/Email.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Value;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
-use Stringable;
 
 use function filter_var;
 
@@ -30,8 +29,8 @@ final readonly class Email
         $this->value = $value;
     }
 
-    public static function fromString(Stringable|string $value): self
+    public static function fromString(string $value): self
     {
-        return new self((string) $value);
+        return new self($value);
     }
 }