+++ src/Firebase/Value/Uid.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Value;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
-use Stringable;
 
 use function mb_strlen;
 
@@ -28,8 +27,8 @@ final readonly class Uid
         $this->value = $value;
     }
 
-    public static function fromString(Stringable|string $value): self
+    public static function fromString(string $value): self
     {
-        return new self((string) $value);
+        return new self($value);
     }
 }